+++ bot/__main__.py
@@ -57,6 +57,7 @@ bot.load_extension("bot.cogs.defcon")
 bot.load_extension("bot.cogs.eval")
 bot.load_extension("bot.cogs.free")
 bot.load_extension("bot.cogs.information")
+bot.load_extension("bot.cogs.infractions")
 bot.load_extension("bot.cogs.jams")
 bot.load_extension("bot.cogs.moderation")
 bot.load_extension("bot.cogs.off_topic_names")
